requirements.txt
asgiref==3.5.2
attrs==21.4.0
autobahn==22.4.2
Automat==20.2.0
backports.zoneinfo==0.2.1
Cerberus==1.3.4
certifi==2022.5.18.1
cffi==1.15.0
channels==3.0.4
charset-normalizer==2.0.12
constantly==15.1.0
cryptography==37.0.2
daphne==3.0.2
defusedxml==0.7.1
Deprecated==1.2.13
Django==3.2.10
django-braces==1.15.0
django-cors-headers==3.12.0
django-oauth-toolkit==2.0.0
django-otp==1.1.3
django-rest-framework-social-oauth2==1.1.0
django-rest-passwordreset==1.2.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
jwcrypto==1.3
oauthlib==3.2.0
Pillow==9.1.1
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.4.0
pyOpenSSL==22.0.0
python-decouple==3.6
python3-openid==3.2.0
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
service-identity==21.1.0
shortuuid==1.0.9
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
Twisted==22.4.0
txaio==22.2.1
typing-extensions==4.2.0
urllib3==1.26.9
wrapt==1.14.1
zope.interface==5.4.0
